requirements.txt
# Core dependencies
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
pyyaml==6.0.1
python-dateutil==2.8.2

# Data collection and APIs
aiohttp==3.9.1
requests==2.31.0
asyncio-throttle==1.0.2
websockets==12.0

# Data processing and analysis
scipy==1.11.4
scikit-learn==1.3.2
statsmodels==0.14.1

# Geospatial
geopy==2.4.1
folium==0.15.1

# Data storage
pyarrow==14.0.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
